Changes to Bibliography and Appendix
Bibliography Style Changes:
This commit adds support for a bibliography style that Felipe Ortega modified from the existing WM Bib style. This modified version allows for the use of hyperrefs via the doi field and no longer enforces the alphabetical supporting of the citations. This bibliography style is called wmbib_nosrt_wurl.bst. I have also added the APSRev4.2 bibliography style, which is common among physics literature. This bibliography style is called apsrev4-2_custom.bst. To select the desired bibliography style one must modify the makeThesisBib command in the wmthesis.cls file. Both of these files would be dependent on the hyperref package for the hyperlinks to be enabled. The APSRev4.2 style is also dependent on the natbib package for proper formatting.

Also included in this commit is a change to the wmthesis.cls to enable the proper display of appendices as "Appendix A" in the Table of Contents, rather than "Chapter A". This modification was implemented by Felipe Ortega and I am integrated it into the repo here.